Try calm down the threading insanity of this crap
Gives JFR own queue where it spawns its threads. Makes the runloop only
run for a 5 seconds so we have opportunity to tell it to stop. (By
setting the atomic isRunLoop to false).

This seems to fix starving when going offline for long period, and not
cause tons of threads.